chore: prepare Release 35 for testing
