+++ meteor/server/migration/currentSystemVersion.ts
@@ -33,7 +33,9 @@
  // Note: The minor version makes a jump here. 32 now refers to "Release 32".
  * 1.32.0: Release 32  (2021-05-04)
  * 1.33.0: Release 33  (2021-05)
+ * 1.34.0: Release 34 (2021-05)
+ * 1.35.0: Release 35 (2021-06)
  */
 
 // Note: Only set this to release versions, (ie X.Y.Z), not pre-releases (ie X.Y.Z-0-pre-release)
-export const CURRENT_SYSTEM_VERSION = '1.34.0'
+export const CURRENT_SYSTEM_VERSION = '1.35.0'
