+++ meteor/server/migration/migrations.ts
@@ -99,6 +99,9 @@ addSteps1_33_0()
 import { addSteps as addSteps1_34_0 } from './1_34_0'
 addSteps1_34_0()
 
+import { addSteps as addSteps1_35_0 } from './1_35_0'
+addSteps1_35_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
